+++ photutils/utils/colormaps.py
@@ -3,38 +3,11 @@
 This module provides tools for generating matplotlib colormaps.
 """
 
-from astropy.utils import deprecated
 import numpy as np
 
 from .check_random_state import check_random_state
 
 __all__ = ['make_random_cmap']
-
-
-@deprecated('0.7', alternative='make_random_cmap')
-def random_cmap(ncolors=256, random_state=None):
-    """
-    Make a matplotlib colormap consisting of (random) muted colors.
-
-    A random colormap is very useful for plotting segmentation images.
-
-    Parameters
-    ----------
-    ncolors : int, optional
-        The number of colors in the colormap.  The default is 256.
-
-    random_state : int or `~numpy.random.mtrand.RandomState`, optional
-        The pseudo-random number generator state used for random
-        sampling.  Separate function calls with the same
-        ``random_state`` will generate the same colormap.
-
-    Returns
-    -------
-    cmap : `matplotlib.colors.ListedColormap`
-        The matplotlib colormap with random colors.
-    """
-
-    return make_random_cmap(ncolors=ncolors, random_state=random_state)
 
 
 def make_random_cmap(ncolors=256, random_state=None):
@@ -62,10 +35,10 @@ def make_random_cmap(ncolors=256, random_state=None):
     from matplotlib import colors
 
     prng = check_random_state(random_state)
-    h = prng.uniform(low=0.0, high=1.0, size=ncolors)
-    s = prng.uniform(low=0.2, high=0.7, size=ncolors)
-    v = prng.uniform(low=0.5, high=1.0, size=ncolors)
-    hsv = np.dstack((h, s, v))
+    hue = prng.uniform(low=0.0, high=1.0, size=ncolors)
+    sat = prng.uniform(low=0.2, high=0.7, size=ncolors)
+    val = prng.uniform(low=0.5, high=1.0, size=ncolors)
+    hsv = np.dstack((hue, sat, val))
     rgb = np.squeeze(colors.hsv_to_rgb(hsv))
 
     return colors.ListedColormap(rgb)